A simple secret role app
Sometimes there are games, activities, or exercises where you need a
group of people to all know one thing except for a specific number of
individuals.

This app will facilitate that.

In it's current form, it will only highlight one individual and does so
by telling all individuals a word except this one.